Refactor

+++ core/src/ru/mikroacse/rolespell/app/model/game/items/weapons/Weapon.java
@@ -29,8 +29,8 @@ public class Weapon extends Item {
         attackTimer = TimerParser.parse(config.get("attack-timer"));
         damage = IntervalParser.parse(config.get("damage"));
 
-        attackDistance = (double) config.get("attack-distance");
-        limit = (int) config.get("limit");
+        attackDistance = config.get("attack-distance");
+        limit = config.get("limit");
     }
 
     public Timer getAttackTimer() {
